+++ src/datachain/toolkit/__init__.py
@@ -1,4 +1,3 @@
-from .diff import compare
 from .split import train_test_split
 
-__all__ = ["compare", "train_test_split"]
+__all__ = ["train_test_split"]
